+++ src/bindgen/mod.rs
@@ -54,6 +54,7 @@ mod rename;
 mod reserved;
 mod utilities;
 mod writer;
+mod language_backend;
 
 #[allow(unused)]
 pub(crate) use self::cargo::*;
